kernel can preemp inside the syscall or in ring0, implemented loop devices, added modules system with a filesystem registry. polling is added as well as sleep with timer integration, truncation. Fized syscall interface for open. Unfotunetly my chmod funckery on my disk for some other thing every file is modified accroding to the git, even though there's no change :/. For userspace my stdlib C is extended, succesfully compile fasm, link with the  single object file, and asemble a x86 assembly file in the disk and execute the resultant binary.

+++ include/vt.h
@@ -44,7 +44,7 @@ struct vt {
 int vt_redraw();
 int vt_tty_send(int scancode);
 int install_virtual_terminals(int count, int row, int cols);
-void vt_console_putchar( unsigned short c);
+void vt_console_putchar( unsigned int c);
 
 #define KDGETMODE	0x4B3B	//Get current mode (KD_TEXT/KD_GRAPHICS)
 #define KDSETMODE	0x4B3A	//Set mode (text/graphics)
